fixed some warnings

+++ modules/core/test/test_mat.cpp
@@ -2699,7 +2699,7 @@ TEST(Mat, reshape_0d)
     EXPECT_EQ(m0d.type(), CV_32S);
     EXPECT_EQ(m0d.ptr<int>(), &v);
     EXPECT_EQ(m0d.rows, 1);
-    EXPECT_EQ(m0d.total(), 1);
+    EXPECT_EQ(m0d.total(), (size_t)1);
     EXPECT_EQ(m0d.empty(), false);
 
     int sz = 1;
@@ -2708,21 +2708,21 @@ TEST(Mat, reshape_0d)
     EXPECT_EQ(m1d.type(), CV_32S);
     EXPECT_EQ(m1d.ptr<int>(), &v);
     EXPECT_EQ(m1d.rows, 1);
-    EXPECT_EQ(m1d.total(), 1);
+    EXPECT_EQ(m1d.total(), (size_t)1);
 
     Mat m2d = m1d.reshape(0, 1);
     EXPECT_EQ(m2d.dims, 2);
     EXPECT_EQ(m2d.type(), CV_32S);
     EXPECT_EQ(m2d.ptr<int>(), &v);
     EXPECT_EQ(m2d.size(), Size(1, 1));
-    EXPECT_EQ(m2d.total(), 1);
+    EXPECT_EQ(m2d.total(), (size_t)1);
 
     Mat m0d_ = m2d.reshape(0, 0, nullptr);
     EXPECT_EQ(m0d_.dims, 0);
     EXPECT_EQ(m0d_.type(), CV_32S);
     EXPECT_EQ(m0d_.ptr<int>(), &v);
     EXPECT_EQ(m0d_.size(), Size(1, 1));
-    EXPECT_EQ(m0d_.total(), 1);
+    EXPECT_EQ(m0d_.total(), (size_t)1);
     EXPECT_EQ(m0d_.empty(), false);
 }
 
@@ -2737,7 +2737,7 @@ TEST(Mat, reshape_empty)
     EXPECT_EQ(m.type(), CV_32S);
     EXPECT_EQ(m.ptr<int>(), nullptr);
     EXPECT_EQ(m.rows, 1);
-    EXPECT_EQ(m.total(), 0);
+    EXPECT_EQ(m.total(), (size_t)0);
 }
